ci: fix packages
