+++ pyproject.toml
@@ -38,6 +38,7 @@ pandas = {version="~1.3", python=">=3.7.1", optional=true}
 apache-flink = {version = "^1.17.0", optional = true}
 ydata-profiling = {version = "^4.2.0", optional = true, python = ">=3.8,<3.12"}
 pyyaml = {version = "^6.0", optional = true}
+pymongo = "^3.8.0"
 
 [tool.poetry.group.test.dependencies]
 pytest = "^7.1.2"
